+++ thunder/time.h
@@ -4,17 +4,17 @@ class Time
 {
 	int x;
 	int y;
-	int totalTime;
-	int timeLeft;
+	size_t totalTime;
+	size_t timeLeft;
 	char timeSymbol;
 	bool isColor = false;
 
 public:
-	const int& getTimeLeft() const { return timeLeft;}
+	const size_t& getTimeLeft() const { return timeLeft;}
 	void setLocation(int x, int y) { this->x = x; this->y = y; }
-	void setTimeSettings(const int totalTime, const bool isColor);
+	void setTimeSettings(const size_t totalTime, const bool isColor);
 	bool checkAndupdateTime();
 	void resetTime() { timeLeft = totalTime; }
-	void printTime(int time);
+	void printTime(size_t time);
 };
